backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.8.0
aiohttp>=3.9.0
pydantic>=2.5.0
python-dotenv>=1.0.0
yfinance>=0.2.31
sgp4>=2.22
numpy>=1.26.0
